encryptly/requirements.txt
pymysql
django
grunt
python-memcached
channels
django-js-reverse
twilio
pycrypto
django-unixdatetimefield
Pillow
asgi_redis
django-fullurl